+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.2.0] - 2026-04-05
+### Added
+- Checking that the type is a structure
+- Checking that the type is a component
+- Initialization and Processing logs
+- Added labels to dll: RunOnlyOnAssembliesWithReference, SourceGenerator
+- Added package dependencies
+
+### Fixed
+- Remove Unity source generation markers such as SystemAPI and IJobEntity
+
 ## [0.1.0] - 2026-04-03
 ### Added
 - .NET project for source generator dll